+++ da-test/src/constants.rs
@@ -1,0 +1,3 @@
+pub const ETHEREUM_DA_CONFIG: &str = include_str!("../../examples/da-confs/ethereum.json");
+pub const CELESTIA_DA_CONFIG: &str = include_str!("../../examples/da-confs/celestia.json");
+pub const AVAIL_DA_CONFIG: &str = include_str!("../../examples/da-confs/avail.json");